// src/app/shared/models/interfaces/response-objects/ro-detailed-person.ts
export interface IRoDetailedPerson {
    adult: boolean
    also_known_as: string []
    biography: string
    birthday: string
    deathday: string
    gender: number
    id: number
    imdb_id: string
    known_for_department: string
    name: string
    place_of_birth: string
    popularity: number
    profile_path: string
}